+++ ESMF/GOCART2G_GridComp/GA_Environment/GA_EnvironmentMod.F90
@@ -17,10 +17,10 @@ module GA_EnvironmentMod
        real, allocatable      :: rhop(:)        ! soil class density [kg m-3]
        real, allocatable      :: fscav(:)       ! scavenging efficiency
 !      logical                :: scav_byColdCloud ! new flag example
-       real, allocatable      :: molwght(:)     ! molecular weight            !NOT UNIVERSAL ONLY FOR GASES, 
+       real, allocatable      :: molwght(:)     ! molecular weight            !NOT UNIVERSAL ONLY FOR GASES,
        real, allocatable      :: fnum(:)        ! number of particles per kg mass
        real, allocatable      :: fwet_ice(:)    ! large scale wet removal scaling factor for ice
-       real, allocatable      :: fwet_snow(:)   ! large scale wet removal scaling factor for snow 
+       real, allocatable      :: fwet_snow(:)   ! large scale wet removal scaling factor for snow
        real, allocatable      :: fwet_rain(:)   ! large scale wet removal scaling factor for rain
        real                   :: washout_tuning ! tuning factor for washout process (1 by default)
        real                   :: wet_radius_thr ! wet radius threshold [um]
@@ -32,6 +32,7 @@ module GA_EnvironmentMod
        real                   :: plid           ! pressure lid [hPa]
        integer                :: klid           ! vertical index of pressure lid
        character(:), allocatable :: wet_removal_scheme     ! name of wet removal scheme
+       character(:), allocatable :: settling_scheme ! settling option (1 - use SettlingSolver, 2 - use SettlingSolverUFS)
        real, allocatable      :: wavelengths_profile(:) ! wavelengths for profile aop [nm]
        real, allocatable      :: wavelengths_vertint(:) ! wavelengths for vertically integrated aop [nm]
     contains
@@ -55,6 +56,7 @@ module GA_EnvironmentMod
 
        !   Local variables
        character(len=ESMF_MAXSTR) :: wet_removal_scheme
+       character(len=ESMF_MAXSTR) :: settling_scheme
 
        !   Get nbins from cfg
        call ESMF_ConfigGetAttribute (cfg, self%nbins, label='nbins:', __RC__)
@@ -70,7 +72,7 @@ module GA_EnvironmentMod
                 self%wavelengths_profile(n_wavelengths_profile), &
                 self%wavelengths_vertint(n_wavelengths_vertint), &
                 __STAT__)
-       
+
        call ESMF_ConfigGetAttribute (cfg, self%radius,     label='particle_radius_microns:', __RC__)
        call ESMF_ConfigGetAttribute (cfg, self%rhop,       label='particle_density:', __RC__)
        call ESMF_ConfigGetAttribute (cfg, self%fscav,      label='fscav:', __RC__)
@@ -83,7 +85,9 @@ module GA_EnvironmentMod
        call ESMF_ConfigGetAttribute (cfg, self%wet_radius_thr,  label='wet_radius_thr:', default=0.05, __RC__)
        call ESMF_ConfigGetAttribute (cfg, self%washout_tuning,  label='washout_tuning:', default=1.0, __RC__)
        call ESMF_ConfigGetAttribute (cfg, wet_removal_scheme, label='wet_removal_scheme:', default='gocart', __RC__)
+       call ESMF_ConfigGetAttribute (cfg, self%settling_scheme, label='settling_scheme:', default='gocart', __RC__)
        self%wet_removal_scheme = ESMF_UtilStringLowerCase(trim(wet_removal_scheme), __RC__)
+       self%settling_scheme = ESMF_UtilStringLowerCase(trim(self%settling_scheme), __RC__)
 
        call ESMF_ConfigGetAttribute (universal_cfg, self%wavelengths_profile, label='wavelengths_for_profile_aop_in_nm:', __RC__)
        call ESMF_ConfigGetAttribute (universal_cfg, self%wavelengths_vertint, &
@@ -116,6 +120,7 @@ module GA_EnvironmentMod
 
        !   * Wet removal scheme
        _ASSERT_RC(any(self%wet_removal_scheme == [character(len=7) :: 'gocart','ufs']), "Error. Unallowed wet removal scheme: "//trim(self%wet_removal_scheme)//". Allowed: gocart, ufs", ESMF_RC_NOT_IMPL)
+       _ASSERT_RC(any(self%settling_scheme == [character(len=7) :: 'gocart','ufs']), "Error. Unallowed settling option: "//trim(self%settling_scheme)//". Allowed: gocart, ufs", ESMF_RC_NOT_IMPL)
 
     end subroutine validate_config