+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Heartbeat time step removed and `timeToWork` logical added
 - Modified filepaths for the optics files to no longer link to a personal nobackup directory
 - Added logic to ensure the alarm accounts for skipping heartbeat in `NI` and `SU` components
+- Added new SettlingSolver scheme `SettlingSolverUFS` and rolled back the previous `SettlingSolver`
 
 ### Fixed
 
@@ -29,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix settling solver for numerical instability under certain convective conditions
 
 ### Added
+
+ - Added option for choosing SettlingSolver
+   - `gocart' - Default GOCART Settling scheme | 'ufs' - New Settling scheme
 
 ## [v2.4.2] - 2025-06-12